fixing CLang C++ compiler errors and warnings
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1224 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/num/polynomial.h
@@ -104,7 +104,7 @@ public:
 
 	Polynomial<T> derivative() const;
 	Polynomial<T> integral() const;
-	Polynomial<T> pow(unsigned iPower) const;
+	Polynomial<T> pow(size_t iPower) const;
 
 	const size_type size() const;
 	const const_iterator begin() const;
